bevy_hanabi
Module attributes

Effect attributes, like the position or velocity of a particle.

An effect attribute is a quantity stored per particle for all particles. Unlike properties, each particle can have a different value for each attribute. Examples of particle attributes include the particle’s own position and its velocity. Attributes are represented by the Attribute type.

Attributes are indirectly added to an effect by adding modifiers requiring them. Each modifier documents its required attributes. You can force a single attribute by adding the SetAttributeModifier.

Note that 🎆 Hanabi provides a number of associated Attribute constants, like Attribute::POSITION. You cannot build your own Attribute instance. See Built-in attributes and Custom attributes for all available attributes.

Definition

Attribute defines the attribute’s unique name, the type of its value, and a default value used to initialize the attribute if not otherwise explicitly initialized.

The attribute name is a string used to identify the attribute, and also as the associated variable name in any WGSL shader code using that attribute. Because it’s unique, attributes can be compared by their name alone.

The attribute type encodes the type of data stored in the attribute, including the number of components for a vector or matrix type. It’s stored together with the default value for the attribute into a Value field.

Layout

Each particle effect contains one or more attributes. The set of all attributes makes a particle. Attributes are organized in memory into a layout, which optimizes GPU RAM usage and access by packing types together, and avoid any padding gaps. However this layout needs to follow the rules of WGSL structs, therefore might introduce some gaps (wasted space) nonetheless, depending on each attribute’s required type alignment.

Here’s an example for the default particle attribute set, containing the particle’s position and velocity vectors (vec3<f32>), and its age and lifetime (f32), packed into a 32 bytes struct:

Bytes	0..4	4..8	8..12	12..16

0..16	POSITION.X	POSITION.Y	POSITION.Z	AGE

16..32	VELOCITY.X	VELOCITY.Y	VELOCITY.Z	LIFETIME

In WGSL code, this is represented by:

struct Particle {
  position: vec3<f32>,
  age: f32,
  velocity: vec3<f32>,
  lifetime: f32,
}

The layout of a particle effect is represented by the ParticleLayout type, and built from a set of attributes via the ParticleLayoutBuilder helper. This is done internally by 🎆 Hanabi for each effect, so in general you don’t have to use those types directly.

Built-in attributes

🎆 Hanabi provides a number of built-in attributes with a specified meaning. Those attributes are interpreted by some built-in library systems in a way specific to the attribute. For example, the Attribute::POSITION represents the particle’s own position, and will be used as the position where to render the particle.

In general those attributes can be read and written by the user, for example via the SetAttributeModifier, but are also read and/or modified by 🎆 Hanabi itself.

Attribute	Meaning

Attribute::POSITION	The particle’s position in simulation space.

Attribute::VELOCITY	The particle’s velocity in simulation space.

Attribute::AGE	The particle’s age, in seconds.

Attribute::LIFETIME	The particle’s total lifetime, in seconds.

Attribute::COLOR	The particle’s LDR color as u32.

Attribute::HDR_COLOR	The particle’s HDR color as vec4<f32>.

Attribute::ALPHA	The particle’s opacity.

Attribute::SIZE	The particle’s uniform size.

Attribute::SIZE2	The particle’s non-uniform size.

Attribute::AXIS_X	X axis of the particle frame.

Attribute::AXIS_Y	Y axis of the particle frame.

Attribute::AXIS_Z	Z axis of the particle frame.

Attribute::SPRITE_INDEX	Index of the current sprite for flipbook animation.

Custom attributes

In additon of the built-in attributes, 🎆 Hanabi provides a number of custom attributes, which are attributes with a specified type but no particular internal meaning. Users are free to use those attributes to store any quantity they like, noting that each new attribute increases the per-particle size and therefor the total RAM usage of the particle effect.

Attribute	Meaning

Attribute::F32_0	A custom f32 attribute.

Attribute::F32_1	A custom f32 attribute.

Attribute::F32_2	A custom f32 attribute.

Attribute::F32_3	A custom f32 attribute.

Attribute::F32X2_0	A custom vec2<f32> attribute.

Attribute::F32X2_1	A custom vec2<f32> attribute.

Attribute::F32X2_2	A custom vec2<f32> attribute.

Attribute::F32X2_3	A custom vec2<f32> attribute.

Attribute::F32X3_0	A custom vec3<f32> attribute.

Attribute::F32X3_1	A custom vec3<f32> attribute.

Attribute::F32X3_2	A custom vec3<f32> attribute.

Attribute::F32X3_3	A custom vec3<f32> attribute.

Attribute::F32X4_0	A custom vec4<f32> attribute.

Attribute::F32X4_1	A custom vec4<f32> attribute.

Attribute::F32X4_2	A custom vec4<f32> attribute.

Attribute::F32X4_3	A custom vec4<f32> attribute.

Structs

Attribute

An attribute of a particle simulated for an effect.

MatrixType

Floating-point matrix type (matCxR<f32>).

ParticleLayout

Particle layout of an effect.

ParticleLayoutBuilder

Builder helper to create a new ParticleLayout.

VectorType

Vector type (vecN<T>).

Enums

ScalarType

Scalar types.

ValueType

Type of an Attribute’s value.