Struct bevy_hanabi::attributes::Attribute

pub struct Attribute(/* private fields */);

An attribute of a particle simulated for an effect.

Effects are composed of many simulated particles. Each particle is in turn composed of a set of attributes, which are used to simulate and render it. Common attributes include the particle’s position, its age, or its color. See Attribute::all() for a list of supported attributes. User-created attributes are not supported.

See also the attributes module documentation for more details about particle attributes.

Implementations

impl Attribute

pub const POSITION: Attribute = _

The particle position in simulation space.

Name

position

Type

VectorType::VEC3F representing the XYZ coordinates of the position.

pub const VELOCITY: Attribute = _

The particle velocity in simulation space.

Name

velocity

Type

VectorType::VEC3F representing the XYZ coordinates of the velocity.

pub const AGE: Attribute = _

The age of the particle.

Each time the particle is updated, the current simulation delta time is added to the particle’s age. The age can be used to animate some other quantities; see the ColorOverLifetimeModifier for example.

If the particle also has a lifetime (either a per-effect constant value, or a per-particle value stored in the Attribute::LIFETIME attribute), then when the age of the particle exceeds its lifetime, the particle dies and is not simulated nor rendered anymore.

Name

age

Type

ScalarType::Float

pub const LIFETIME: Attribute = _

The lifetime of the particle.

This attribute stores a per-particle lifetime, which compared to the particle’s age allows determining if the particle needs to be simulated and rendered. This requires the Attribute::AGE attribute to be used too.

Name

lifetime

Type

ScalarType::Float

pub const COLOR: Attribute = _

The particle’s base color.

This attribute stores a per-particle color, which can be used for various purposes, generally as the base color for rendering the particle.

Name

color

Type

ScalarType::Uint representing the RGBA components of the color encoded as 0xAABBGGRR, with a single byte per component, where the alpha value is stored in the most significant byte and the red value in the least significant byte. Note that this representation is the same as the one returned by Color::as_rgba_u32() and Color::as_linear_rgba_u32().

pub const HDR_COLOR: Attribute = _

The particle’s base color (HDR).

This attribute stores a per-particle HDR color, which can be used for various purposes, generally as the base color for rendering the particle.

Name

hdr_color

Type

VectorType::VEC4F representing the RGBA components of the color. Values are not clamped, and can be outside the [0:1] range to represent HDR values.

pub const ALPHA: Attribute = _

The particle’s opacity (alpha).

This is a value in [0:1], where 0 corresponds to a fully transparent particle, and 1 to a fully opaque one.

Name

alpha

Type

ScalarType::Float

pub const SIZE: Attribute = _

The particle’s uniform size.

The particle is uniformly scaled by this size.

Name

size

Type

ScalarType::Float

pub const SIZE2: Attribute = _

The particle’s 2D size, for quad rendering.

The particle, when drawn as a quad, is scaled along its local X and Y axes by these values.

Name

size2

Type

VectorType::VEC2F representing the XY sizes of the particle.

pub const PREV: Attribute = _

The previous particle in the ribbon chain.

Name

prev

Type

ScalarType::Uint representing the index of the previous particle in the chain.

pub const NEXT: Attribute = _

The next particle in the ribbon chain.

Name

next

Type

ScalarType::Uint representing the index of the next particle in the chain.

pub const AXIS_X: Attribute = _

The local X axis of the particle.

This attribute stores a per-particle X axis, which defines the horizontal direction of a quad particle. This is generally used to re-orient the particle during rendering, for example to face the camera or another point of interest. For example, the OrientModifier modifies this attribute to make the particle face a specific item.

Name

axis_x

Type

VectorType::VEC3F

pub const AXIS_Y: Attribute = _

The local Y axis of the particle.

This attribute stores a per-particle Y axis, which defines the vertical direction of a quad particle. This is generally used to re-orient the particle during rendering, for example to face the camera or another point of interest. For example, the OrientModifier modifies this attribute to make the particle face a specific item.

Name

axis_y

Type

VectorType::VEC3F

pub const AXIS_Z: Attribute = _

The local Z axis of the particle.

This attribute stores a per-particle Z axis, which defines the normal to a quad particle’s plane. This is generally used to re-orient the particle during rendering, for example to face the camera or another point of interest. For example, the OrientModifier modifies this attribute to make the particle face a specific item.

Name

axis_z

Type

VectorType::VEC3F

pub const SPRITE_INDEX: Attribute = _

The sprite index in a flipbook animation.

This attribute stores the index of the sprite of a flipbook animation. This is used with the FlipbookModifier.

Name

sprite_index

Type

ScalarType::Int

pub const F32_0: Attribute = _

A generic scalar float attribute.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32_0

Type

ScalarType::Float

pub const F32_1: Attribute = _

A generic scalar float attribute.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32_1

Type

ScalarType::Float

pub const F32_2: Attribute = _

A generic scalar float attribute.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32_2

Type

ScalarType::Float

pub const F32_3: Attribute = _

A generic scalar float attribute.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32_3

Type

ScalarType::Float

pub const F32X2_0: Attribute = _

A generic vector float attribute with 2 components.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32x2_0

Type

VectorType::VEC2F

pub const F32X2_1: Attribute = _

A generic vector float attribute with 2 components.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32x2_1

Type

VectorType::VEC2F

pub const F32X2_2: Attribute = _

A generic vector float attribute with 2 components.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32x2_2

Type

VectorType::VEC2F

pub const F32X2_3: Attribute = _

A generic vector float attribute with 2 components.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32x2_3

Type

VectorType::VEC2F

pub const F32X3_0: Attribute = _

A generic vector float attribute with 3 components.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32x3_0

Type

VectorType::VEC3F

pub const F32X3_1: Attribute = _

A generic vector float attribute with 3 components.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32x3_1

Type

VectorType::VEC3F

pub const F32X3_2: Attribute = _

A generic vector float attribute with 3 components.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32x3_2

Type

VectorType::VEC3F

pub const F32X3_3: Attribute = _

A generic vector float attribute with 3 components.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32x3_3

Type

VectorType::VEC3F

pub const F32X4_0: Attribute = _

A generic vector float attribute with 4 components.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32x4_0

Type

VectorType::VEC4F

pub const F32X4_1: Attribute = _

A generic vector float attribute with 4 components.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32x4_1

Type

VectorType::VEC4F

pub const F32X4_2: Attribute = _

A generic vector float attribute with 4 components.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32x4_2

Type

VectorType::VEC4F

pub const F32X4_3: Attribute = _

A generic vector float attribute with 4 components.

This attribute can be used for anything. It has no specific meaning. You can store whatever per-particle value you want in it (for example, at spawn time) and read it back later.

Name

f32x4_3

Type

VectorType::VEC4F

pub fn from_name(name: &str) -> Option<Attribute>

Retrieve an attribute by its name.

See Attribute::all() for the list of attributes, and the Attribute::name() method of each attribute for their name.

Example

let attr = Attribute::from_name("position").unwrap();
assert_eq!(attr, Attribute::POSITION);

pub fn all() -> &'static [Attribute]

Get the list of all existing attributes.

Example

for attr in Attribute::all() {
    println!("{}", attr.name());
}

pub fn name(&self) -> &'static str

The attribute’s name.

The name of an attribute is unique, and corresponds to the name of the variable in the generated WGSL code.

pub fn default_value(&self) -> Value

The attribute’s default value.

pub fn value_type(&self) -> ValueType

The attribute’s type.

This is a shortcut for default_value().value_type().

pub fn size(&self) -> usize

Size of this attribute, in bytes.

This is a shortcut for value_type().size().

pub fn align(&self) -> usize

Alignment of this attribute, in bytes.

This is a shortcut for value_type().align().