Function bevy_hanabi::tick_spawners

pub fn tick_spawners(
    commands: Commands<'_, '_>,
    time: Res<'_, Time<EffectSimulation>>,
    effects: Res<'_, Assets<EffectAsset>>,
    rng: ResMut<'_, Random>,
    query: Query<'_, '_, (Entity, &ParticleEffect, Option<&InheritedVisibility>, Option<&mut EffectSpawner>)>,
)

Tick all the EffectSpawner components of the simulated ParticleEffect components.

This system runs in the PostUpdate stage, after the visibility system has updated the InheritedVisibility of each effect instance (see VisibilitySystems::VisibilityPropagate). Hidden instances are not updated, unless the EffectAsset::simulation_condition is set to SimulationCondition::Always.

Note that by that point the ViewVisibility is not yet calculated, and it may happen that spawners are ticked but no effect is visible in any view even though some are “visible” (active) in the World. The actual per-view culling of invisible (not in view) effects is performed later on the render world.

Once the system determined that the effect instance needs to be simulated this frame, it ticks the spawner by calling EffectSpawner::tick(), adding a new EffectSpawner component if it doesn’t already exist on the same entity as the ParticleEffect.