bevy_hanabi
Module graph

Effect graph and language definition.

This module contains the elements used to build an effect graph, a fully customizable description of a visual effect.

The effect graph API is composed of two layers:

The Expression API provides a micro-language to build shader expressions (Expr) via code. Those expressions are composed together into complex behaviors assigned to the input values of some modifiers. This enables complete customizing of the modifiers. This API is focused on runtime execution and asset serialization. It provides the user with a way to indirectly write effect shaders without any knowledge about shaders, and with a framework which guarantees the shader code generated is correct.

The Node API provides a higher-level API built on top of the Expression API. Its use is entirely optional. It defines a node graph (Graph), where each node (Node) represents an expression or a modifier. Nodes are linked together to implicitly build expressions. This API focuses on asset editing, with the explicit intent to be used to build a (visual) effect editor.

API status

Currently effect graphs are not fully available yet; only some preview elements exist.

The Expression API already contains a good set of expressions, and some modifiers have already been converted to accept expressions for their input fields. Its generally in a reasonable shape for early adoption.

The Node API contains a basic node and graph definition, which is entirely experimental at this stage.

We recommend starting to familiarize yourself with effect graphs, and starting to port your code to use expressions (Expr), as the entire library is moving in the direction of adopting effect graphs across the board.

Re-exports

pub use expr::AttributeExpr;

pub use expr::BinaryOperator;

pub use expr::BuiltInExpr;

pub use expr::BuiltInOperator;

pub use expr::EvalContext;

pub use expr::Expr;

pub use expr::ExprError;

pub use expr::ExprHandle;

pub use expr::ExprWriter;

pub use expr::LiteralExpr;

pub use expr::Module;

pub use expr::PropertyExpr;

pub use expr::UnaryOperator;

pub use expr::WriterExpr;

pub use node::AddNode;

pub use node::AttributeNode;

pub use node::DivNode;

pub use node::Graph;

pub use node::MulNode;

pub use node::Node;

pub use node::NormalizeNode;

pub use node::Slot;

pub use node::SlotDir;

pub use node::SlotId;

pub use node::SubNode;

pub use node::TimeNode;

Modules

expr

Expression API

node

Node API.

Structs

MatrixValue

Floating-point matrix value.

VectorValue

Variant storage for a vector value.

Enums

ScalarValue

Variant storage for a single (scalar) value.

ScalarValueMut

Variant storage for a scalar value.

Value

Variant storage for a simple value.

Traits

ElemType

Trait to convert the elements of a vector.