Struct bevy_hanabi::graph::expr::WriterExpr

pub struct WriterExpr { /* private fields */ }

Intermediate expression from an ExprWriter.

A writer expression WriterExpr is equivalent to an ExprHandle, but retains a reference to the underlying Module and therefore can easily be chained with other WriterExpr via a concise syntax, at the expense of being more heavyweight and locking the underlying Module under a ref-counted interior mutability (Rc<RefCell<Module>>). ExprHandle by opposition is a very lightweight type, similar to a simple index. And like an array index, ExprHandle doesn’t explicitly reference its associated storage (Module) which needs to be remembered by the user explicitly.

In addition, WriterExpr implements several numerical operators like the std::ops::Add trait, making it simpler to combine it with another WriterExpr.

let mut w = ExprWriter::new();
let x = w.lit(-3.5);
let y = w.lit(78.);
let z = x + y; // == 74.5

In general the WriterExpr type is not used directly, but inferred from calling ExprWriter methods and combining WriterExpr together.

let mut w = ExprWriter::new();
let my_prop = w.add_property("my_prop", 3.0.into());

// x = max(-3.5 + 1., properties.my_prop) * 0.5 - particle.position;
let x = (w.lit(-3.5) + w.lit(1.)).max(w.prop(my_prop)).mul(w.lit(0.5))
    .sub(w.attr(Attribute::POSITION));

let handle: ExprHandle = x.expr();

Implementations

impl WriterExpr

pub fn abs(self) -> Self

Take the absolute value of the current expression.

This is a unary operator, which applies component-wise to vector and matrix operand expressions.

Example

// A literal expression `x = -3.5;`.
let x = w.lit(-3.5);

// The absolute value `y = abs(x);`.
let y = x.abs(); // == 3.5

pub fn all(self) -> Self

Apply the logical operator “all” to the current bool vector expression.

This is a unary operator, which applies to vector operand expressions to produce a scalar boolean.

Example

// A literal expression `x = vec3<bool>(true, false, true);`.
let x = w.lit(BVec3::new(true, false, true));

// Check if all components are true `y = all(x);`.
let y = x.all(); // == false

pub fn any(self) -> Self

Apply the logical operator “any” to the current bool vector expression.

This is a unary operator, which applies to vector operand expressions to produce a scalar boolean.

Example

// A literal expression `x = vec3<bool>(true, false, true);`.
let x = w.lit(BVec3::new(true, false, true));

// Check if any components is true `y = any(x);`.
let y = x.any(); // == true

pub fn ceil(self) -> Self

Apply the “ceil” operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Ceil: `y = ceil(x);`
let y = x.ceil();

pub fn cos(self) -> Self

Apply the “cos” operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Cos: `y = cos(x);`
let y = x.cos();

pub fn exp(self) -> Self

Apply the “exp” operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Exp: `y = exp(x);`
let y = x.exp();

pub fn exp2(self) -> Self

Apply the “exp2” operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Exp2: `y = exp2(x);`
let y = x.exp2();

pub fn floor(self) -> Self

Apply the “floor” operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Floor: `y = floor(x);`
let y = x.floor();

pub fn fract(self) -> Self

Apply the “fract” operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Fract: `y = fract(x);`
let y = x.fract();

pub fn inverse_sqrt(self) -> Self

Apply the “inverseSqrt” (inverse square root) operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Inverse square root: `y = inverseSqrt(x) = 1.0 / sqrt(x);`
let y = x.inverse_sqrt();

pub fn length(self) -> Self

Apply the “length” operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Length: `y = length(x);`
let y = x.length();

pub fn log(self) -> Self

Apply the “log” operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Log: `y = log(x);`
let y = x.log();

pub fn log2(self) -> Self

Apply the “log2” operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Log2: `y = log2(x);`
let y = x.log2();

pub fn normalized(self) -> Self

Apply the “normalize” operator to the current float vector expression.

This is a unary operator, which applies to float vector operand expressions to produce another float vector with unit length (normalized).

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Normalize: `y = normalize(x);`
let y = x.normalized();

pub fn pack4x8snorm(self) -> Self

Apply the “pack4x8snorm” operator to the current 4-component float vector expression.

This is a unary operator, which applies to 4-component float vector operand expressions to produce a single u32 scalar expression.

Example

// A literal expression `x = vec4<f32>(-1., 1., 0., 7.2);`.
let x = w.lit(Vec4::new(-1., 1., 0., 7.2));

// Pack: `y = pack4x8snorm(x);`
let y = x.pack4x8snorm(); // 0x7F007FFFu32

pub fn pack4x8unorm(self) -> Self

Apply the “pack4x8unorm” operator to the current 4-component float vector expression.

This is a unary operator, which applies to 4-component float vector operand expressions to produce a single u32 scalar expression.

Example

// A literal expression `x = vec4<f32>(-1., 1., 0., 7.2);`.
let x = w.lit(Vec4::new(-1., 1., 0., 7.2));

// Pack: `y = pack4x8unorm(x);`
let y = x.pack4x8unorm(); // 0xFF00FF00u32

pub fn sign(self) -> Self

Apply the “sign” operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Sign: `y = sign(x);`
let y = x.sign();

pub fn sin(self) -> Self

Apply the “sin” operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Sin: `y = sin(x);`
let y = x.sin();

pub fn sqrt(self) -> Self

Apply the “sqrt” (square root) operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Square root: `y = sqrt(x);`
let y = x.sqrt();

pub fn tan(self) -> Self

Apply the “tan” operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Tan: `y = tan(x);`
let y = x.tan();

pub fn unpack4x8snorm(self) -> Self

Apply the “unpack4x8snorm” operator to the current u32 scalar expression.

This is a unary operator, which applies to u32 scalar operand expressions to produce a 4-component floating point vector of signed normalized components in [-1:1].

Example

// A literal expression `y = 0x7F007FFFu32;`.
let y = w.lit(0x7F007FFFu32);

// Unpack: `x = unpack4x8snorm(y);`
let x = y.unpack4x8snorm(); // vec4<f32>(-1., 1., 0., 7.2)

pub fn unpack4x8unorm(self) -> Self

Apply the “unpack4x8unorm” operator to the current u32 scalar expression.

This is a unary operator, which applies to u32 scalar operand expressions to produce a 4-component floating point vector of unsigned normalized components in [0:1].

Example

// A literal expression `y = 0xFF00FF00u32;`.
let y = w.lit(0xFF00FF00u32);

// Unpack: `x = unpack4x8unorm(y);`
let x = y.unpack4x8unorm(); // vec4<f32>(-1., 1., 0., 7.2)

pub fn saturate(self) -> Self

Apply the “saturate” operator to the current float scalar or vector expression.

This is a unary operator, which applies to float scalar or vector operand expressions to produce a float scalar or vector. It applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(1., 1., 1.);`.
let x = w.lit(Vec3::ONE);

// Saturate: `y = saturate(x);`
let y = x.saturate();

pub fn x(self) -> Self

Get the first component of a scalar or vector.

Example

// A literal expression `v = vec3<f32>(1., 1., 1.);`.
let v = w.lit(Vec3::ONE);

// f = v.x;`
let f = v.x();

pub fn y(self) -> Self

Get the second component of a vector.

Example

// A literal expression `v = vec3<f32>(1., 1., 1.);`.
let v = w.lit(Vec3::ONE);

// f = v.y;`
let f = v.y();

pub fn z(self) -> Self

Get the third component of a vector.

Example

// A literal expression `v = vec3<f32>(1., 1., 1.);`.
let v = w.lit(Vec3::ONE);

// f = v.z;`
let f = v.z();

pub fn w(self) -> Self

Get the fourth component of a vector.

Example

// A literal expression `v = vec3<f32>(1., 1., 1.);`.
let v = w.lit(Vec3::ONE);

// f = v.w;`
let f = v.w();

pub fn add(self, other: Self) -> Self

Add the current expression with another expression.

This is a binary operator, which applies component-wise to vector operand expressions.

You can also use the std::ops::Add trait directly, via the + symbol, as an alternative to calling this method directly.

Example

// A literal expression `x = vec2<f32>(3., -2.);`.
let x = w.lit(Vec2::new(3., -2.));

// Another literal expression `y = vec2<f32>(1., 5.);`.
let y = w.lit(Vec2::new(1., 5.));

// The sum of both vectors `z = x + y;`.
let z = x.add(y); // == vec2<f32>(4., 3.)
// -OR-
// let z = x + y;

pub fn cross(self, other: Self) -> Self

Calculate the cross product of the current expression by another expression.

This is a binary operator, which applies to vector operands of size 3 only, and always produces a vector of the same size.

Example

// A literal expression `x = vec3<f32>(3., -2., 1.);`.
let x = w.lit(Vec3::new(3., -2., 1.));

// Another literal expression `y = vec3<f32>(1., 5., 0.);`.
let y = w.lit(Vec3::new(1., 5., 0.));

// The cross product of both vectors `z = cross(x, y);`.
let z = x.cross(y);

pub fn dot(self, other: Self) -> Self

Calculate the dot product of the current expression by another expression.

This is a binary operator, which applies to vector operands of same size only, and always produces a floating point scalar.

Example

// A literal expression `x = vec2<f32>(3., -2.);`.
let x = w.lit(Vec2::new(3., -2.));

// Another literal expression `y = vec2<f32>(1., 5.);`.
let y = w.lit(Vec2::new(1., 5.));

// The dot product of both vectors `z = dot(x, y);`.
let z = x.dot(y);

pub fn distance(self, other: Self) -> Self

Calculate the distance between the current expression and another expression.

This is a binary operator.

Example

// A literal expression `x = vec3<f32>(3., -2., 1.);`.
let x = w.lit(Vec3::new(3., -2., 1.));

// Another literal expression `y = vec3<f32>(1., 5., 0.);`.
let y = w.lit(Vec3::new(1., 5., 0.));

// The distance between the vectors `z = distance(x, y);`.
let z = x.distance(y);

pub fn div(self, other: Self) -> Self

Divide the current expression by another expression.

This is a binary operator, which applies component-wise to vector operand expressions.

You can also use the std::ops::Div trait directly, via the / symbol, as an alternative to calling this method directly.

Example

// A literal expression `x = vec2<f32>(3., -2.);`.
let x = w.lit(Vec2::new(3., -2.));

// Another literal expression `y = vec2<f32>(1., 5.);`.
let y = w.lit(Vec2::new(1., 5.));

// The quotient of both vectors `z = x / y;`.
let z = x.div(y); // == vec2<f32>(3., -0.4)
// -OR-
// let z = x / y;

pub fn ge(self, other: Self) -> Self

Apply the logical operator “greater than or equal” to this expression and another expression.

This is a binary operator, which applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(3., -2., 7.);`.
let x = w.lit(Vec3::new(3., -2., 7.));

// Another literal expression `y = vec3<f32>(1., 5., 7.);`.
let y = w.lit(Vec3::new(1., 5., 7.));

// The boolean result of the greater than or equal operation `z = (x >= y);`.
let z = x.ge(y); // == vec3<bool>(true, false, true)

pub fn gt(self, other: Self) -> Self

Apply the logical operator “greater than” to this expression and another expression.

This is a binary operator, which applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(3., -2., 7.);`.
let x = w.lit(Vec3::new(3., -2., 7.));

// Another literal expression `y = vec3<f32>(1., 5., 7.);`.
let y = w.lit(Vec3::new(1., 5., 7.));

// The boolean result of the greater than operation `z = (x > y);`.
let z = x.gt(y); // == vec3<bool>(true, false, false)

pub fn le(self, other: Self) -> Self

Apply the logical operator “less than or equal” to this expression and another expression.

This is a binary operator, which applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(3., -2., 7.);`.
let x = w.lit(Vec3::new(3., -2., 7.));

// Another literal expression `y = vec3<f32>(1., 5., 7.);`.
let y = w.lit(Vec3::new(1., 5., 7.));

// The boolean result of the less than or equal operation `z = (x <= y);`.
let z = x.le(y); // == vec3<bool>(false, true, true)

pub fn lt(self, other: Self) -> Self

Apply the logical operator “less than” to this expression and another expression.

This is a binary operator, which applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(3., -2., 7.);`.
let x = w.lit(Vec3::new(3., -2., 7.));

// Another literal expression `y = vec3<f32>(1., 5., 7.);`.
let y = w.lit(Vec3::new(1., 5., 7.));

// The boolean result of the less than operation `z = (x < y);`.
let z = x.lt(y); // == vec3<bool>(false, true, false)

pub fn max(self, other: Self) -> Self

Take the maximum value of the current expression and another expression.

This is a binary operator, which applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec2<f32>(3., -2.);`.
let x = w.lit(Vec2::new(3., -2.));

// Another literal expression `y = vec2<f32>(1., 5.);`.
let y = w.lit(Vec2::new(1., 5.));

// The maximum of both vectors `z = max(x, y);`.
let z = x.max(y); // == vec2<f32>(3., 5.)

pub fn min(self, other: Self) -> Self

Take the minimum value of the current expression and another expression.

This is a binary operator, which applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec2<f32>(3., -2.);`.
let x = w.lit(Vec2::new(3., -2.));

// Another literal expression `y = vec2<f32>(1., 5.);`.
let y = w.lit(Vec2::new(1., 5.));

// The minimum of both vectors `z = min(x, y);`.
let z = x.min(y); // == vec2<f32>(1., -2.)

pub fn mul(self, other: Self) -> Self

Multiply the current expression with another expression.

This is a binary operator, which applies component-wise to vector operand expressions.

You can also use the std::ops::Mul trait directly, via the * symbol, as an alternative to calling this method directly.

Example

// A literal expression `x = vec2<f32>(3., -2.);`.
let x = w.lit(Vec2::new(3., -2.));

// Another literal expression `y = vec2<f32>(1., 5.);`.
let y = w.lit(Vec2::new(1., 5.));

// The product of both vectors `z = x * y;`.
let z = x.mul(y); // == vec2<f32>(3., -10.)
// -OR-
// let z = x * y;

pub fn rem(self, other: Self) -> Self

Calculate the remainder of the division of the current expression by another expression.

This is a binary operator.

Example

// A literal expression `x = vec3<f32>(3., -2., 1.);`.
let x = w.lit(Vec3::new(3., -2., 1.));

// Another literal expression `y = vec3<f32>(1., 5., 0.);`.
let y = w.lit(Vec3::new(1., 5., 0.));

// The remainder of the division `z = x % y;`.
let z = x.rem(y);

pub fn step(self, edge: Self) -> Self

Calculate the step of a value with respect to a reference.

This is a binary operator, which applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec3<f32>(3., -2.);`.
let x = w.lit(Vec3::new(3., -2., 8.));

// An edge reference `e = vec3<f32>(1., 5.);`.
let e = w.lit(Vec3::new(1., 5., 8.));

// The step value
let s = x.step(e); // == vec3<f32>(1., 0., 1.)

pub fn sub(self, other: Self) -> Self

Subtract another expression from the current expression.

This is a binary operator, which applies component-wise to vector operand expressions.

You can also use the std::ops::Sub trait directly, via the - symbol, as an alternative to calling this method directly.

Example

// A literal expression `x = vec2<f32>(3., -2.);`.
let x = w.lit(Vec2::new(3., -2.));

// Another literal expression `y = vec2<f32>(1., 5.);`.
let y = w.lit(Vec2::new(1., 5.));

// The difference of both vectors `z = x - y;`.
let z = x.sub(y); // == vec2<f32>(2., -7.)
// -OR-
// let z = x - y;

pub fn uniform(self, other: Self) -> Self

Apply the logical operator “uniform” to this expression and another expression.

This is a binary operator, which applies component-wise to vector operand expressions. That is, for vectors, this produces a vector of random values where each component is uniformly distributed within the bounds of the related component of both operands.

Example

// A literal expression `x = vec3<f32>(3., -2., 7.);`.
let x = w.lit(Vec3::new(3., -2., 7.));

// Another literal expression `y = vec3<f32>(1., 5., 7.);`.
let y = w.lit(Vec3::new(1., 5., 7.));

// A random variable uniformly distributed in [1:3]x[-2:5]x[7:7].
let z = x.uniform(y);

pub fn mix(self, other: Self, fraction: Self) -> Self

Blending linearly (“mix”) two expressions with the fraction provided by a third expression.

This is a ternary operator, which applies component-wise to vector operand expressions.

Example

// A literal expression `x = vec2<f32>(3., -2.);`.
let x = w.lit(Vec2::new(3., -2.));

// Another literal expression `y = vec2<f32>(1., 4.);`.
let y = w.lit(Vec2::new(1., 4.));

// A fraction `t = 0.5;`.
let t = w.lit(0.25);

// The linear blend of x and y via t: z = (1 - t) * x + y * t
let z = x.mix(y, t); // == vec2<f32>(2.5, -0.5)

pub fn smoothstep(self, low: Self, high: Self) -> Self

Calculate the smooth Hermite interpolation in [0:1] of the current value taken between the given bounds.

This is a ternary operator, which applies component-wise to vector operand expressions.

Example

// A literal expression `low = vec2<f32>(3., -2.);`.
let low = w.lit(Vec2::new(3., -2.));

// Another literal expression `high = vec2<f32>(1., 4.);`.
let high = w.lit(Vec2::new(1., 4.));

// A point `x = vec2<f32>(2., 1.);` between `low` and `high`.
let x = w.lit(Vec2::new(2., 1.));

// The smooth Hermite interpolation: `t = smoothstep(low, high, x)`
let t = x.smoothstep(low, high); // == 0.5

pub fn vec2(self, y: Self) -> Self

Construct a Vec2 from two scalars.

Example

let theta = w.add_property("theta", 0.0.into());
// Convert the angular property `theta` to a 2D vector.
let (cos_theta, sin_theta) = (w.prop(theta).cos(), w.prop(theta).sin());
let circle_pos = cos_theta.vec2(sin_theta);

pub fn vec3(self, y: Self, z: Self) -> Self

Construct a Vec3 from two scalars.

Example

let theta = w.add_property("theta", 0.0.into());
// Convert the angular property `theta` to a 3D vector in a flat plane.
let (cos_theta, sin_theta) = (w.prop(theta).cos(), w.prop(theta).sin());
let circle_pos = cos_theta.vec3(w.lit(0.0), sin_theta);

pub fn cast(self, target: impl Into<ValueType>) -> Self

Cast an expression to a different type.

Example

let x = w.lit(Vec2::new(3., -2.));
let y = x.cast(VectorType::VEC3I); // x = vec3<i32>(particle.position);

pub fn expr(self) -> ExprHandle

Finalize an expression chain and return the accumulated expression.

The returned handle indexes the Module owned by the ExprWriter this intermediate expression was built from.

Example

// A literal expression `x = -3.5;`.
let x = w.lit(-3.5);

// Retrieve the ExprHandle for that expression.
let handle = x.expr();