bevy_hanabi::graph::expr
Trait EvalContext

pub trait EvalContext {
    // Required methods
    fn modifier_context(&self) -> ModifierContext;
    fn particle_layout(&self) -> &ParticleLayout;
    fn property_layout(&self) -> &PropertyLayout;
    fn eval(
        &mut self,
        module: &Module,
        handle: ExprHandle,
    ) -> Result<String, ExprError>;
    fn make_local_var(&mut self) -> String;
    fn push_stmt(&mut self, stmt: &str);
    fn make_fn(
        &mut self,
        func_name: &str,
        args: &str,
        module: &mut Module,
        f: &mut dyn FnMut(&mut Module, &mut dyn EvalContext) -> Result<String, ExprError>,
    ) -> Result<(), ExprError>;
    fn is_attribute_pointer(&self) -> bool;
}

Evaluation context for transforming expressions into WGSL code.

The evaluation context references a Module storing all Expr in use, as well as a ParticleLayout defining the existing attributes of each particle and their layout in memory, and a PropertyLayout defining existing properties and their layout in memory. These together define the context within which expressions are evaluated.

A same expression can be valid in one context and invalid in another. The most obvious example is a PropertyExpr which is only valid if the property is actually defined in the property layout of the evaluation context.

Required Methods

fn modifier_context(&self) -> ModifierContext

Get the modifier context of the evaluation.

fn particle_layout(&self) -> &ParticleLayout

Get the particle layout of the effect.

fn property_layout(&self) -> &PropertyLayout

Get the property layout of the effect.

fn eval( &mut self, module: &Module, handle: ExprHandle, ) -> Result<String, ExprError>

Evaluate an expression, returning its WGSL shader code.

The evaluation is guaranteed to be unique. Calling eval() multiple times with the same handle will return the same result. If the expression referenced by handle has side effects, it’s evaluated only once on first call, stored in a local variable, and the variable cached and returned on subsequent calls.

fn make_local_var(&mut self) -> String

Generate a unique local variable name.

Each time this function is called, a new unique name is generated. The name is guaranteed to be unique within the current evaluation context only. Do not use for global top-level identifiers.

The variable name is not registered automatically in the Module. If you call make_local_var() but doesn’t use the returned name, it won’t appear in the shader.

fn push_stmt(&mut self, stmt: &str)

Push an intermediate statement during an evaluation.

Intermediate statements are inserted before the expression evaluation which produced them. They’re generally used to define temporary local variables, for example to store the result of expressions with side effects.

fn make_fn( &mut self, func_name: &str, args: &str, module: &mut Module, f: &mut dyn FnMut(&mut Module, &mut dyn EvalContext) -> Result<String, ExprError>, ) -> Result<(), ExprError>

Create a function.

Create a new function with the given func_name inside the given Module. The function takes a list of arguments args, which are copied verbatim into the shader code without any validation. The body of the function is generated by invoking the given closure once with the input module and a temporary EvalContext local to the function. The closure must return the generated shader code of the function body. Any statement pushed to the temporary function context with EvalContext::push_stmt() is emitted inside the function body before the returned code. The function can subsequently be called from the parent context by generating code to call func_name, with the correct arguments.

fn is_attribute_pointer(&self) -> bool

Check if the particle attribute struct is a pointer?

In some context the attribute struct (named ‘particle’ in WGSL code) is a pointer instead of being a struct instance. This happens in particular when defining a function for a modifier, and passing the attribute struct to be modified. In that case the generated code needs to emit a pointer indirection code to access the fields of the struct.