Module bevy_hanabi::graph::node

Node API.

⚠️ This API is under construction and is missing some features. ⚠️

The Node API is designed for effect editing, and the creation of UI tools. It builds on top of the Expression API and is entirely optional. It defines a Graph composed of Nodes. Each node represent either a Modifier or an Expr. A node has some Slots associated with it, representing its inputs and outputs, if any. An output slot of a node can be linked to an input slot of another node to express that the value produced by the upstream node must flow through to the input of the downstream node.

An effect Graph can be serialized as is, to retain its editing capabilities. Alternatively, once the user has finished building an effect, it can be converted to a runtime EffectAsset for use as a ParticleEffect.

Structs

AddNode

Graph node to add two values.

AttributeNode

Graph node to get any single particle attribute.

DivNode

Graph node to divide two values.

Graph

Effect graph.

MulNode

Graph node to multiply two values.

NodeId

Identifier of a node in a graph.

NormalizeNode

Graph node to normalize a vector value.

Slot

Single slot of a node.

SlotDef

Definition of a slot of a node.

SlotId

Identifier of a slot in a graph.

SubNode

Graph node to subtract two values.

TimeNode

Graph node to get various time values related to the effect system.

Enums

SlotDir

Node slot direction.

Traits

Node

Generic graph node.