bevy_hanabi::graph
Trait ElemType

pub trait ElemType {
    const ZERO: Self;

    // Required methods
    fn get_n<const N: usize>(storage: &[u32; 4]) -> Self;
    fn get(index: usize, storage: &[u32; 4]) -> Self;
    fn get_all(storage: &[u32; 4], count: usize) -> &[Self]
       where Self: Sized;
    fn get_mut(index: usize, storage: &mut [u32; 4]) -> &mut Self;
}

Trait to convert the elements of a vector.

Required Associated Constants

const ZERO: Self

The zero element of the type.

Required Methods

fn get_n<const N: usize>(storage: &[u32; 4]) -> Self

Get the N-th component of the vector given its raw storage.

fn get(index: usize, storage: &[u32; 4]) -> Self

Get the given component of the vector given its raw storage.

fn get_all(storage: &[u32; 4], count: usize) -> &[Self]
where Self: Sized,

Get a slice of all the components of the vector given its raw storage.

This is only valid for numeric types, and will panic for a boolean type.

fn get_mut(index: usize, storage: &mut [u32; 4]) -> &mut Self

Get a mutable reference to the given component of the vector from within its raw storage.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.