Module bevy_hanabi::modifier::force

Modifiers to apply forces to the particles.

The forces are applied as accelerations times unit mass, as particles currently do not have a mass.

Structs

ConformToSphereModifier

A modifier to apply a force to the particle which makes it conform (“stick”) to the surface of a sphere.

LinearDragModifier

A modifier to apply a linear drag force to all particles each frame. The force slows down the particles without changing their direction.