bevy_hanabi::modifier
Module output

Modifiers to influence the output (rendering) of each particle.

Structs

ColorOverLifetimeModifier

A modifier modulating each particle’s color over its lifetime with a gradient curve.

FlipbookModifier

A modifier to render particles using flipbook animation.

OrientModifier

Orients the particle’s local frame.

ParticleTextureModifier

A modifier modulating each particle’s color by sampling a texture.

RoundModifier

Makes particles round.

ScreenSpaceSizeModifier

A modifier to interpret the size of all particles in screen-space pixels.

SetColorModifier

A modifier to set the rendering color of all particles.

SetSizeModifier

A modifier to set the size of all particles.

SizeOverLifetimeModifier

A modifier modulating each particle’s size over its lifetime with a gradient curve.

Enums

ImageSampleMapping

Mapping of the sample read from a texture image to the base particle color.

OrientMode

Mode of orientation of a particle’s local frame.