Module bevy_hanabi::modifier::position

Modifiers to set the position of particles.

These modifiers directly manipulate a particle’s position. They’re generally useful to initialize the position at spawn time, but can occasionally be used during simulation update to enforce a particular position.

Structs

SetPositionCircleModifier

A modifier to set the position of particles on or inside a circle/disc, randomly.

SetPositionCone3dModifier

A modifier to set the position of particles on a truncated 3D cone.

SetPositionSphereModifier

A modifier to set the position of particles on or inside a sphere, randomly.