Struct bevy_hanabi::modifier::RenderContext

pub struct RenderContext<'a> {
    pub property_layout: &'a PropertyLayout,
    pub particle_layout: &'a ParticleLayout,
    pub vertex_code: String,
    pub fragment_code: String,
    pub render_extra: String,
    pub particle_texture: Option<Handle<Image>>,
    pub image_sample_mapping_code: String,
    pub sprite_grid_size: Option<UVec2>,
    pub gradients: HashMap<u64, Gradient<Vec4>>,
    pub size_gradients: HashMap<u64, Gradient<Vec2>>,
    pub needs_uv: bool,
    /* private fields */
}

Particle rendering shader code generation context.

Fields
property_layout: &'a PropertyLayout

Layout of properties for the current effect.

particle_layout: &'a ParticleLayout

Layout of attributes of a particle for the current effect.

vertex_code: String

Main particle rendering code for the vertex shader.

fragment_code: String

Main particle rendering code for the fragment shader.

render_extra: String

Extra functions emitted at top level, which vertex_code and fragment_code can call.

particle_texture: Option<Handle<Image>>

Texture modulating the particle color.

image_sample_mapping_code: String

WGSL code describing how to modulate the base color of the particle with the image texture sample, if any.

sprite_grid_size: Option<UVec2>

Flipbook sprite sheet grid size, if any.

gradients: HashMap<u64, Gradient<Vec4>>

Color gradients.

size_gradients: HashMap<u64, Gradient<Vec2>>

Size gradients.

needs_uv: bool

Needs uv

Implementations

impl<'a> RenderContext<'a>

pub fn new( property_layout: &'a PropertyLayout, particle_layout: &'a ParticleLayout, ) -> Self

Create a new update context.

pub fn with_attribute_pointer(self) -> Self

Mark the attribute struct as being available through a pointer.