bevy_hanabi::modifier
Module velocity

Modifiers to set the velocity of particles.

These modifiers directly manipulate a particle’s velocity. They’re generally useful to initialize the velocity at spawn time, but can occasionally be used during simulation update to enforce a particular velocity. The velocity in turn drives the position update during Euler motion integration.

particle.position += particle.velocity * simulation.delta_time;

Structs

SetVelocityCircleModifier

A modifier to set the velocity of particles radially on a circle.

SetVelocitySphereModifier

A modifier to set the velocity of particles to a spherical distribution.

SetVelocityTangentModifier

A modifier to set the velocity of particles along the tangent to an axis.