Struct bevy_hanabi::EffectAsset

pub struct EffectAsset {
    pub name: String,
    pub spawner: Spawner,
    pub z_layer_2d: f32,
    pub simulation_space: SimulationSpace,
    pub simulation_condition: SimulationCondition,
    pub motion_integration: MotionIntegration,
    pub alpha_mode: AlphaMode,
    /* private fields */
}

Asset describing a visual effect.

The effect can be instanciated with a ParticleEffect component, or a ParticleEffectBundle.

Fields
name: String

Display name of the effect.

This has no internal use, and is mostly for the user to identify an effect or for display in some tool UI. It’s however used in serializing the asset.

spawner: Spawner

Spawner.

z_layer_2d: f32

For 2D rendering, the Z coordinate used as the sort key.

This value is passed to the render pipeline and used when sorting transparent items to render, to order them. As a result, effects with different Z values cannot be batched together, which may negatively affect performance.

Ignored for 3D rendering.

simulation_space: SimulationSpace

Particle simulation space.

simulation_condition: SimulationCondition

Condition under which the effect is simulated.

motion_integration: MotionIntegration

Type of motion integration applied to the particles of a system.

alpha_mode: AlphaMode

Alpha mode.

Implementations

impl EffectAsset

pub fn new(capacities: Vec<u32>, spawner: Spawner, module: Module) -> Self

Create a new effect asset.

The effect assets requires 2 essential pieces:

The capacities of the effect, which together represent the maximum number of particles which can be stored and simulated at the same time for each group. There will be one capacity value per group; thus, the capacities array also specifies the number of groups. All capacities must be non-zero and should be the smallest possible values which allow you to author the effect. These values directly impact the GPU memory consumption of the effect, which will allocate some buffers to store that many particles for as long as the effect exists. The capacities of an effect are immutable. See also capacities() for more details.

The Spawner, which defines when particles are emitted. All spawners spawn particles into group 0. (To add particles to other groups, use the crate::modifier::clone::CloneModifier.)

Additionally, if any modifier added to this effect uses some Expr to customize its behavior, then those Expr are stored into a Module which should be passed to this method. If expressions are not used, just pass an empty module Module::default().

Examples

Create a new effect asset without any modifier. This effect doesn’t really do anything because e.g. the particles have a zero lifetime.

let spawner = Spawner::rate(5_f32.into()); // 5 particles per second
let module = Module::default();
let effect = EffectAsset::new(vec![32768], spawner, module);

Create a new effect asset with a modifier holding an expression. The expression is stored inside the Module transfered to the EffectAsset.

let spawner = Spawner::rate(5_f32.into()); // 5 particles per second

let mut module = Module::default();

// Create a modifier that initialized the particle lifetime to 10 seconds.
let lifetime = module.lit(10.); // literal value "10.0"
let init_lifetime = SetAttributeModifier::new(Attribute::LIFETIME, lifetime);

let effect = EffectAsset::new(vec![32768], spawner, module);

pub fn capacities(&self) -> &[u32]

Get the capacities of the effect, in number of particles per group.

For example, if this function returns &[256, 512], then this effect has two groups, the first of which has a maximum of 256 particles and the second of which has a maximum of 512 particles.

Each value in the array represents the number of particles stored in GPU memory at all time for the group with the corresponding index, even if unused, so you should try to minimize this value. However, the Spawner cannot emit more particles than the capacity of group 0. Whatever the spawner settings, if the number of particles reaches the capacity, no new particle can be emitted. Setting an appropriate capacity for an effect is therefore a compromise between more particles available for visuals and more GPU memory usage.

Common values range from 256 or less for smaller effects, to several hundreds of thousands for unique effects consuming a large portion of the GPU memory budget. Hanabi has been tested with over a million particles, however the performance will largely depend on the actual GPU hardware and available memory, so authors are encouraged not to go too crazy with the capacities.

pub fn module(&self) -> &Module

Get the expression module storing all expressions in use by modifiers of this effect.

pub fn with_name(self, name: impl Into<String>) -> Self

Set the effect name.

The effect name is used when serializing the effect.

pub fn with_simulation_condition( self, simulation_condition: SimulationCondition ) -> Self

Set the effect’s simulation condition.

pub fn with_simulation_space(self, simulation_space: SimulationSpace) -> Self

Set the effect’s simulation space.

pub fn with_alpha_mode(self, alpha_mode: AlphaMode) -> Self

Set the alpha mode.

pub fn properties(&self) -> &[Property]

Get the list of existing properties.

This is a shortcut for self.module().properties().

pub fn init<M>(self, modifier: M) -> Self
where M: Modifier + Send + Sync,

Add an initialization modifier to the effect.

Initialization modifiers only apply to particles that are freshly spawned. Currently, spawners can only spawn into group 0. Consequently, the initialization modifiers will only affect particles in group 0.

Panics

Panics if the modifier doesn’t support the init context (that is, modifier.context() returns a flag which doesn’t include ModifierContext::Init).

pub fn update<M>(self, modifier: M) -> Self
where M: Modifier + Send + Sync,

Add an update modifier to the effect.

Panics

Panics if the modifier doesn’t support the update context (that is, modifier.context() returns a flag which doesn’t include ModifierContext::Update).

pub fn update_groups<M>(self, modifier: M, groups: ParticleGroupSet) -> Self
where M: Modifier + Send + Sync,

Add an update modifier to the effect targeting only a subset of groups.

pub fn add_modifier( self, context: ModifierContext, modifier: Box<dyn Modifier> ) -> Self

Add a [BoxedModifier] to the specific context.

Panics

Panics if the context is ModifierContext::Render; use add_render_modifier() instead.

Panics if the input context contains more than one context (the bitfield contains more than 1 bit set) or no context at all (zero bit set).

Panics if the modifier doesn’t support the context specified (that is, modifier.context() returns a flag which doesn’t include context).

pub fn add_modifier_to_groups( self, context: ModifierContext, modifier: Box<dyn Modifier>, groups: ParticleGroupSet ) -> Self

Add a [BoxedModifier] to the specific context, in a specific set of groups.

Panics

Panics if the context is ModifierContext::Render; use add_render_modifier() instead.

Panics if the input context contains more than one context (the bitfield contains more than 1 bit set) or no context at all (zero bit set).

Panics if the modifier doesn’t support the context specified (that is, modifier.context() returns a flag which doesn’t include context).

pub fn render<M>(self, modifier: M) -> Self
where M: RenderModifier + Send + Sync,

Add a render modifier to the effect.

Panics

Panics if the modifier doesn’t support the render context (that is, modifier.context() returns a flag which doesn’t include ModifierContext::Render).

pub fn render_groups<M>(self, modifier: M, groups: ParticleGroupSet) -> Self
where M: RenderModifier + Send + Sync,

Add a render modifier to specific groups of this effect.

Panics

Panics if the modifier doesn’t support the render context (that is, modifier.context() returns a flag which doesn’t include ModifierContext::Render).

pub fn add_render_modifier(self, modifier: Box<dyn RenderModifier>) -> Self

Add a RenderModifier to the render context.

Panics

Panics if the modifier doesn’t support the render context (that is, modifier.context() returns a flag which doesn’t include ModifierContext::Render).

pub fn add_render_modifier_to_groups( self, modifier: Box<dyn RenderModifier>, groups: ParticleGroupSet ) -> Self

Add a RenderModifier to the render context targeting a specific set of groups.

Panics

Panics if the modifier doesn’t support the render context (that is, modifier.context() returns a flag which doesn’t include ModifierContext::Render).

pub fn modifiers(&self) -> impl Iterator<Item = &dyn Modifier>

Get a list of all the modifiers of this effect.

pub fn init_modifiers(&self) -> impl Iterator<Item = &dyn Modifier>

Get a list of all the init modifiers of this effect.

This is a filtered list of all modifiers, retaining only modifiers executing in the ModifierContext::Init context.

pub fn update_modifiers(&self) -> impl Iterator<Item = &dyn Modifier>

Get a list of all the update modifiers of this effect.

This is a filtered list of all modifiers, retaining only modifiers executing in the ModifierContext::Update context.

pub fn update_modifiers_for_group( &self, group_index: u32 ) -> impl Iterator<Item = &dyn Modifier>

Get a list of all the update modifiers in a single group.

This is a filtered list of all modifiers, retaining only modifiers executing in the ModifierContext::Update context and affecting the specified group.

pub fn render_modifiers(&self) -> impl Iterator<Item = &dyn RenderModifier>

Get a list of all the render modifiers of this effect.

This is a filtered list of all modifiers, retaining only modifiers executing in the ModifierContext::Render context.

pub fn render_modifiers_for_group( &self, group_index: u32 ) -> impl Iterator<Item = &dyn RenderModifier>

Get a list of all the render modifiers of this effect that affect a specific group.

This is a filtered list of all modifiers, retaining only modifiers executing in the ModifierContext::Render context and that affect the given group.

pub fn particle_layout(&self) -> ParticleLayout

Build the particle layout of the asset based on its modifiers.

This method calculates the particle layout of the effect based on the currently existing particles, and return it as a newly allocated ParticleLayout object.

pub fn property_layout(&self) -> PropertyLayout

Build the property layout of the asset based on its properties.

This method calculates the property layout of the effect based on the currently existing properties, and return it as a newly allocated PropertyLayout object.