Trait bevy_hanabi::graph::node::Node

pub trait Node {
    // Required methods
    fn slots(&self) -> &[SlotDef];
    fn eval(
        &self,
        module: &mut Module,
        inputs: Vec<ExprHandle>
    ) -> Result<Vec<ExprHandle>, ExprError>;
}

Generic graph node.

Required Methods

fn slots(&self) -> &[SlotDef]

Get the list of slots of this node.

The list contains both input and output slots, without any guaranteed order.

fn eval( &self, module: &mut Module, inputs: Vec<ExprHandle> ) -> Result<Vec<ExprHandle>, ExprError>

Evaluate the node from the given input expressions, and optionally produce output expression(s).

The expressions themselves are not evaluated (that is, e.g. “3 + 2” is not reduced to “5”).