Module bevy_hanabi::modifier

Building blocks to create a visual effect.

A modifier is a building block used to create effects. Particles effects are composed of multiple modifiers, which put together and configured produce the desired visual effect. Each modifier changes a specific part of the behavior of an effect. Modifiers are grouped in three categories:

Init modifiers influence the initializing of particles when they spawn. They typically configure the initial position and/or velocity of particles. Init modifiers implement the InitModifier trait.

Update modifiers influence the particle update loop each frame. For example, an update modifier can apply a gravity force to all particles. Update modifiers implement the UpdateModifier trait.

Render modifiers influence the rendering of each particle. They can change the particle’s color, or orient it to face the camera. Render modifiers implement the RenderModifier trait.

A single modifier can be part of multiple categories. For example, the SetAttributeModifier can be used either to initialize a particle’s attribute on spawning, or to assign a value to that attribute each frame during simulation (update).

Re-exports

pub use accel::*;

pub use attr::*;

pub use clone::*;

pub use force::*;

pub use kill::*;

pub use output::*;

pub use position::*;

pub use ribbon::*;

pub use velocity::*;

Modules

accel

Modifiers influencing the acceleration of particles.

attr

Modifiers to manipulate a particle’s attributes.

clone

Modifiers to duplicate particles.

force

Modifiers to apply forces to the particles.

kill

Modifiers to kill particles under specific conditions.

output

Modifiers to influence the output (rendering) of each particle.

position

Modifiers to set the position of particles.

ribbon

Renders particles as ribbons.

velocity

Modifiers to set the velocity of particles.

Structs

GroupedModifier

A Modifier that affects to one or more groups.

ModifierContext

Context a modifier applies to.

ParticleGroupSet

A bitfield that describes which particle groups a modifier affects.

RenderContext

Particle rendering shader code generation context.

ShaderWriter

Shader code writer.

Enums

ShapeDimension

The dimension of a shape to consider.

Traits

Modifier

Trait describing a modifier customizing an effect pipeline.

RenderModifier

Trait to customize the rendering of alive particles each frame.

Type Aliases

BoxedModifier

Boxed version of Modifier.