Trait bevy_hanabi::EffectSimulationTime

pub trait EffectSimulationTime {
    // Required methods
    fn relative_speed(&self) -> f32;
    fn relative_speed_f64(&self) -> f64;
    fn effective_speed(&self) -> f32;
    fn effective_speed_f64(&self) -> f64;
    fn set_relative_speed(&mut self, ratio: f32);
    fn set_relative_speed_f64(&mut self, ratio: f64);
    fn pause(&mut self);
    fn unpause(&mut self);
    fn is_paused(&self) -> bool;
    fn was_paused(&self) -> bool;
}

All methods for the Time<EffectSimulation> clock.

Required Methods

fn relative_speed(&self) -> f32

Returns the speed the clock advances relative to the virtual clock, as f32.

fn relative_speed_f64(&self) -> f64

Returns the speed the clock advances relative to the virtual clock, as f64.

fn effective_speed(&self) -> f32

Returns the speed the clock advanced relative to your system clock in this update, as f32.

Returns 0.0 if either the Time<Virtual> or the Time<EffectSimulationTime> was paused and otherwise the product of the relative_speed of the clocks at the start of the update.

fn effective_speed_f64(&self) -> f64

Returns the speed the clock advanced relative to your system clock in this update, as f64.

Returns 0.0 if either the Time<Virtual> or the Time<EffectSimulationTime> was paused and otherwise the product of the relative_speed of the clocks at the start of the update.

fn set_relative_speed(&mut self, ratio: f32)

Sets the speed the clock advances relative to the virtual clock, given as an f32.

For example, setting this to 2.0 will make the clock advance twice as fast as the virtual clock.

Panics

Panics if ratio is negative or not finite.

fn set_relative_speed_f64(&mut self, ratio: f64)

Sets the speed the clock advances relative to the virtual clock, given as an f64.

For example, setting this to 2.0 will make the clock advance twice as fast as the virtual clock.

Panics

Panics if ratio is negative or not finite.

fn pause(&mut self)

Stops the clock, preventing it from advancing until resumed.

fn unpause(&mut self)

Resumes the clock if paused.

fn is_paused(&self) -> bool

Returns true if the clock is currently paused.

fn was_paused(&self) -> bool

Returns true if the clock was paused at the start of this update.